pyproject.toml
[tool.poetry]
name = "disk_shop"
version = "0.1.0"
description = "Backend for game disc store"
authors = ["SailorLekalo <shamshev-andrei@yandex.ru>"]
readme = "README.md"
packages = [{ include = "app" }]


[tool.poetry.dependencies]
python = ">=3.10"
fastapi = ">=0.104.1"
strawberry-graphql = { version = ">=0.212.1", extras = ["fastapi"] }
sqlalchemy = { version = ">=2.0.23", extras = ["asyncio"] }
asyncpg = ">=0.29.0"
alembic = ">=1.13.1"
passlib = { version = ">=1.7.4", extras = ["bcrypt"] }
python-dotenv = ">=1.0.0"
uvicorn = { version = ">=0.23.2", extras = ["standard"] }
python-multipart = ">=0.0.6"
pydantic-settings = ">=2.0.3"
aiogram = ">=3.22.0,<4.0.0"
bcrypt = ">=4.0.1"
psycopg = ">=3.2.9,<4.0.0"
websockets = "^15.0.1"
broadcaster = "^0.3.1"

[tool.poetry.group.dev.dependencies]
ipython = ">=8.17.2"
pytest = ">=7.4.0"
pytest-asyncio = ">=0.23.0"
httpx = ">=0.25.0"
ruff = ">=0.6.0"
pre-commit = ">=3.7.0"

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "uvicorn app.main:app --host 0.0.0.0 --port 8000"
migrate = "alembic upgrade head"
makemigrations = "alembic revision --autogenerate -m"


[tool.alembic]
script_location = "app/migrations"
sqlalchemy.url = "postgresql+asyncpg://${DB_USER}:${DB_PASSWORD}@${DB_HOST}:5432/${DB_NAME}"

[tool.ruff]

src = ["app"]
target-version = "py311"
line-length = 88
exclude = [
    ".venv",
    "migrations",
    "tests"
]
select = ["ALL"]
ignore = [
    "D203",
    "D212",
    "E501",
    "D100", "D101", "D102", "D103", "D104", # docstring не нужны тестовому, полагаю
    "RUF001",
    "UP017",
    "UP035",
    "DTZ005"

]
fix = true
